refactor: refactor item list on native
