+++ packages/author-profile/src/styles/index.js
@@ -4,6 +4,16 @@ import sharedStyles from "./shared";
 
 const styles = StyleSheet.create({
   ...sharedStyles,
+  listItemContainer: {
+    paddingBottom: spacing(3),
+    paddingLeft: spacing(2),
+    paddingRight: spacing(2),
+    paddingTop: spacing(3)
+  },
+  listItemSeparatorContainer: {
+    paddingLeft: spacing(2),
+    paddingRight: spacing(2)
+  },
   pageErrorContainer: {
     flex: 1,
     margin: spacing(2)
@@ -13,6 +23,10 @@ const styles = StyleSheet.create({
     marginBottom: spacing(4),
     marginTop: spacing(4),
     width: "60%"
+  },
+  listErrorContainer: {
+    flex: 1,
+    margin: spacing(3)
   }
 });
